les coordonnées sont des entiers pas des réels donc j'ai change double par int

+++ point.cpp
@@ -3,33 +3,33 @@
 #include <math.h>
 
 using namespace std;
-point::point(double x, double y) : d_x{x} , d_y{y}
+point::point(int x, int y) : d_x{x} , d_y{y}
 {
 }
 point::point() : d_x{0}, d_y{0}
 {
 }
-double point::x() const
+int point::x() const
 {
     return d_x;
 }
-double point::y() const
+int point::y() const
 {
     return d_y;
 }
 double point::distance(point &p)
 {
-    double x = abs(p.x() - d_x);
-    double y = abs(p.y() - d_y);
+    int x = abs(p.x() - d_x);
+    int y = abs(p.y() - d_y);
     double dist = sqrt(x*x + y*y);
     return dist;
 }
-void point::deplaceEn(double dx, double dy)
+void point::deplaceEn(int dx, int dy)
 {
     d_x = dx;
     d_y = dy;
 }
-void point::deplaceDe(double dx, double dy)
+void point::deplaceDe(int dx, int dy)
 {
     d_x += dx;
     d_y += dy;
